// types/site.ts
export type PointerCompanionConfig = {
  label: string;
  showLabel: string;
  hideLabel: string;
  compactLabel: string;
  hint: string;
  catControlLabel: string;
  storageKey: string;
  defaultEnabled: boolean;
  catSrc: string;
  fishSrc: string;
};

export type ClipboardControlCopy = {
  controlLabel: string;
  hintLabel: string;
  successLabel: string;
  failureLabel: string;
  successAnnouncement: string;
  failureAnnouncement: string;
};